requirements.txt
streamlit==1.52.2
langchain==1.2.2
langchain-openai==1.1.7
python-dotenv==1.2.1
pytest==9.0.2
pytest-mock==3.15.1
